requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
executing==1.2.0
idna==3.4
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
pipreqs==0.4.12
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
upgrade-requirements==1.7.0
urllib3==1.26.15
wcwidth==0.2.6
yarg==0.1.9
